Rewrote logic to be alittle clearer

+++ reflections/views.py
@@ -102,9 +102,7 @@ def reflections(request, reflectionId = None):
 	return wrongMethod()
 
 
-def fetchUser(userId):
-	return fetchObject(User, userId)
-
+def fetchUser(userId): return fetchObject(User, userId)
 def createUserIfPossible(request):
 	payload = request.POST
 	required = ['username', 'email', 'password',]
@@ -131,15 +129,16 @@ def deleteUserIfPossible(userId):
 @csrf_exempt 
 def users(request, userId = None):
 	# Se nao for post, retorne erro
-	if not userId == None:
-		if not request.method == 'DELETE': return wrongMethod()
-		
-		return deleteUserIfPossible(userId)
+	if userId == None:
+		if request.method == 'POST': return createUserIfPossible(request)
+		return wrongMethod()
+
+	if request.method == 'DELETE': return deleteUserIfPossible(userId)
+	
+	return wrongMethod()
+	
 
 
 
-
-	if not request.method == 'POST'	: return wrongMethod()
-	return createUserIfPossible(request)
